requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
mysql-connector-python==9.0.0
python-dotenv==1.0.1
APScheduler==3.10.4
firebase-admin==6.5.0
gunicorn==20.1.0
